refactor: unify xmpp data into chatDataStoreSlice
