+++ src/chats/components/CounterBadgeUpdater.tsx
@@ -9,7 +9,7 @@ import { useEffect } from 'react';
 import { updatePrimaryBadge } from '@zextras/carbonio-shell-ui';
 
 import { CHATS_ROUTE } from '../../constants/appConstants';
-import { getTotalUnreadCountSelector } from '../../store/selectors/UnreadsCounterSelectors';
+import { getTotalUnreadCountSelector } from '../../store/selectors/ChatDataSelectors';
 import useStore from '../../store/Store';
 
 const CounterBadgeUpdater = (): null => {
