+++ src/chats/components/conversation/footer/MessageArea.tsx
@@ -109,11 +109,8 @@ const MessageArea: React.FC<MessageAreaPros> = ({
 
 	// Focus input when roomId changes: but if there are unread messages wait to calculate the first unread message
 	useEffect(() => {
-		if (
-			!useStore.getState().unreads[roomId] ||
-			useStore.getState().unreads[roomId] === 0 ||
-			firstNewMessage
-		) {
+		const unread = useStore.getState().chatData[roomId]?.unread;
+		if (!unread || unread === 0 || firstNewMessage) {
 			textareaRef.current?.focus();
 		}
 	}, [firstNewMessage, roomId, textareaRef]);
