+++ src/chats/components/conversation/messageBubbles/readByPopoverList/ReadByPopoverList.tsx
@@ -8,8 +8,7 @@ import React, { ReactElement, useMemo } from 'react';
 import { find, forEach } from 'lodash';
 import { useTranslation } from 'react-i18next';
 
-import { getMarkers } from '../../../../../store/selectors/MarkersSelectors';
-import { getMessagesSelector } from '../../../../../store/selectors/MessagesSelectors';
+import { getMessagesSelector, getMarkers } from '../../../../../store/selectors/ChatDataSelectors';
 import { getUserId } from '../../../../../store/selectors/SessionSelectors';
 import useStore from '../../../../../store/Store';
 import { MessageType } from '../../../../../types/store/MessageTypes';
