+++ src/chats/components/infoPanel/conversationActionsAccordion/ActionsAccordion.tsx
@@ -17,7 +17,7 @@ import EditConversationAction from './EditConversationAction';
 import LeaveConversationAction from './LeaveConversationAction';
 import MuteConversationAction from './MuteConversationAction';
 import { getActionsAccordionStatus } from '../../../../store/selectors/ActiveConversationsSelectors';
-import { roomIsEmpty } from '../../../../store/selectors/MessagesSelectors';
+import { roomIsEmpty } from '../../../../store/selectors/ChatDataSelectors';
 import {
 	getOwnershipOfTheRoom,
 	getNumberOfOwnersOfTheRoom,
