+++ src/meetings/components/bubblesWrapper/MeetingBubble.tsx
@@ -15,11 +15,11 @@ import ForwardInfo from '../../../chats/components/conversation/messageBubbles/F
 import RepliedTextMessageSectionView from '../../../chats/components/conversation/messageBubbles/RepliedTextMessageSectionView';
 import TextContentBubble from '../../../chats/components/conversation/messageBubbles/TextContentBubble';
 import { getInputHasFocus } from '../../../store/selectors/ActiveConversationsSelectors';
-import { getRoomIdByMeetingId } from '../../../store/selectors/MeetingSelectors';
 import {
 	getMessageAttachment,
 	getMessageSelector
-} from '../../../store/selectors/MessagesSelectors';
+} from '../../../store/selectors/ChatDataSelectors';
+import { getRoomIdByMeetingId } from '../../../store/selectors/MeetingSelectors';
 import { getRoomTypeSelector } from '../../../store/selectors/RoomsSelectors';
 import useStore from '../../../store/Store';
 import { MeetingChatVisibility } from '../../../types/store/ActiveMeetingTypes';
