+++ src/store/selectors/ChatDataSelectors.ts
@@ -1,0 +1,227 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Zextras <https://www.zextras.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+import { filter, find, forEach, includes, last, map, orderBy, reduce, size } from 'lodash';
+
+import { FilteredConversation } from '../../chats/components/secondaryBar/SecondaryBarView';
+import { Marker, RoomMarkers } from '../../types/store/MarkersTypes';
+import {
+	AttachmentMessageType,
+	ConfigurationMessage,
+	FasteningAction,
+	Message,
+	MessageFastening,
+	MessageType,
+	TextMessage
+} from '../../types/store/MessageTypes';
+import { RoomType } from '../../types/store/RoomTypes';
+import { RootStore } from '../../types/store/StoreTypes';
+import useStore from '../Store';
+
+export const useOrderedOneToOneAndGroupsInfoByLastMessage = (): FilteredConversation[] => {
+	const { rooms, chatData } = useStore((store) => store);
+	const filteredRooms = filter(
+		rooms,
+		(room) => room.type === RoomType.GROUP || room.type === RoomType.ONE_TO_ONE
+	);
+	const listOfConvLastMessage: FilteredConversation[] = [];
+	forEach(filteredRooms, (room) => {
+		const messages = chatData[room.id]?.messages;
+		const lastMessage = messages && messages[messages.length - 1];
+		listOfConvLastMessage.push({
+			roomId: room.id,
+			name: room.name ?? '',
+			roomType: room.type,
+			lastMessageTimestamp: lastMessage ? lastMessage.date : 0,
+			members: room.members ?? []
+		});
+	});
+	return orderBy(listOfConvLastMessage, ['lastMessageTimestamp'], ['desc']);
+};
+
+const FALLBACK_MESSAGE_SELECTOR: Message[] = [];
+
+export const getMessagesSelector = (store: RootStore, roomId: string): Message[] =>
+	store.chatData[roomId]?.messages ? store.chatData[roomId]?.messages : FALLBACK_MESSAGE_SELECTOR;
+
+const readableMessages: (TextMessage | ConfigurationMessage)[] = [];
+
+export const getReadableMessagesSelector = (
+	store: RootStore,
+	roomId: string
+): (TextMessage | ConfigurationMessage)[] => {
+	readableMessages.length = 0;
+	readableMessages.push(
+		...(filter(store.chatData[roomId]?.messages, (message) =>
+			includes([MessageType.TEXT_MSG, MessageType.CONFIGURATION_MSG], message.type)
+		) as TextMessage[])
+	);
+	return readableMessages;
+};
+
+export const getLastTextMessageIdSelector = (
+	store: RootStore,
+	roomId: string
+): string | undefined => {
+	const textMessages = filter(
+		store.chatData[roomId]?.messages,
+		(message) => message.type === MessageType.TEXT_MSG
+	);
+	if (textMessages && textMessages[textMessages.length - 1]) {
+		return textMessages[textMessages.length - 1].id;
+	}
+	return undefined;
+};
+
+export const getLastMessageIdSelector = (store: RootStore, roomId: string): string | undefined => {
+	const messages = store.chatData[roomId]?.messages;
+	if (messages?.[messages.length - 1]) {
+		return messages[messages.length - 1].id;
+	}
+	return undefined;
+};
+
+export const getMessageSelector = (
+	store: RootStore,
+	roomId: string,
+	messageId: string | undefined
+): Message | undefined =>
+	find(store.chatData[roomId]?.messages, (message) => message.id === messageId);
+
+const listOfConvByLastMessage: {
+	roomId: string;
+	roomType: string;
+	lastMessageTimestamp: number;
+}[] = [];
+
+export const getRoomIdsWithLastMessage = (
+	store: RootStore
+): { roomId: string; roomType: string; lastMessageTimestamp: number }[] => {
+	listOfConvByLastMessage.length = 0;
+	// check to remove and tell BE to improve because if a user is removed from a room
+	// the messages of this always came back and trigger error
+	forEach(store.rooms, (room) => {
+		const messages = store.chatData[room.id]?.messages;
+		const lastMessage = messages?.[messages.length - 1];
+		listOfConvByLastMessage.push({
+			roomId: room.id,
+			roomType: room.type,
+			lastMessageTimestamp: lastMessage ? lastMessage.date : 0
+		});
+	});
+	return listOfConvByLastMessage;
+};
+
+export const roomIsEmpty = (store: RootStore, roomId: string): boolean =>
+	size(store.chatData[roomId]?.messages) === 0;
+
+export const getMessageAttachment = (
+	store: RootStore,
+	message: Message | undefined
+): AttachmentMessageType | undefined => {
+	if (message?.type === MessageType.TEXT_MSG) {
+		const textMessage = find(
+			store.chatData[message.roomId]?.messages,
+			(mex) => mex.id === message.id
+		) as TextMessage;
+		return textMessage?.attachment;
+	}
+	return undefined;
+};
+
+export const getEditAndDeleteFasteningSelector = (
+	state: RootStore,
+	roomId: string,
+	stanzaId: string
+): MessageFastening | undefined => {
+	if (state.chatData[roomId]?.fastenings?.[stanzaId]) {
+		const editAndDeleteFastenings = state.chatData[roomId]?.fastenings?.[stanzaId].filter(
+			(fastening) =>
+				fastening.action === FasteningAction.EDIT || fastening.action === FasteningAction.DELETE
+		);
+		return last(editAndDeleteFastenings);
+	}
+	return undefined;
+};
+
+const filteredReactions: MessageFastening[] = [];
+
+export const getReactionFastenings = (
+	state: RootStore,
+	roomId: string,
+	stanzaId: string
+): MessageFastening[] => {
+	filteredReactions.length = 0;
+	const fastenings = state.chatData[roomId]?.fastenings?.[stanzaId];
+	if (fastenings) {
+		const reactions = filter(
+			fastenings,
+			(fastening) => fastening.action === FasteningAction.REACTION
+		);
+		const latestReactions = reduce(
+			reactions,
+			(acc: { [from: string]: number }, fastening: MessageFastening, index: number) => {
+				if (!acc[fastening.from] || reactions[acc[fastening.from]].date < fastening.date) {
+					acc[fastening.from] = index;
+				}
+				return acc;
+			},
+			{}
+		);
+		filteredReactions.push(
+			...filter(reactions, (_, index) => includes(Object.values(latestReactions), index))
+		);
+		return filteredReactions;
+	}
+	return filteredReactions;
+};
+
+export const getMyLastReaction = (
+	state: RootStore,
+	roomId: string,
+	stanzaId: string
+): string | undefined => {
+	const fastenings = state.chatData[roomId]?.fastenings?.[stanzaId];
+	if (fastenings) {
+		const myReactions = filter(
+			fastenings,
+			(fastening) =>
+				fastening.action === FasteningAction.REACTION && fastening.from === state.session?.id
+		);
+		return last(myReactions)?.value;
+	}
+	return undefined;
+};
+
+export const getMyLastMarkerOfRoom = (store: RootStore, roomId: string): Marker | null => {
+	if (store.session.id && store.chatData[roomId]?.markers[store.session.id]) {
+		return store.chatData[roomId]?.markers[store.session.id];
+	}
+	return null;
+};
+
+export const getRoomHasMarkers = (store: RootStore, roomId: string): boolean =>
+	!!store.chatData[roomId]?.markers;
+
+export const getMarkers = (store: RootStore, roomId: string): RoomMarkers =>
+	store.chatData[roomId]?.markers;
+
+export const getTotalUnreadCountSelector = (store: RootStore): number => {
+	const sum = (amount: number, n: number): number => amount + n;
+	return reduce(
+		map(store.chatData, ({ unread }, key) => {
+			const room = store.rooms[key];
+			if (!!room && !room.userSettings?.muted && room.type !== 'temporary') {
+				return unread;
+			}
+			return 0;
+		}),
+		sum,
+		0
+	);
+};
+
+export const getRoomUnreadsSelector = (store: RootStore, roomId: string): number =>
+	store.chatData[roomId]?.unread || 0;
