--- src/store/selectors/FasteningsSelectors.ts
@@ -1,74 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-
-import { filter, includes, last, reduce } from 'lodash';
-
-import { FasteningAction, MessageFastening } from '../../types/store/MessageTypes';
-import { RootStore } from '../../types/store/StoreTypes';
-
-export const getEditAndDeleteFasteningSelector = (
-	state: RootStore,
-	roomId: string,
-	stanzaId: string
-): MessageFastening | undefined => {
-	if (state.fastenings?.[roomId]?.[stanzaId]) {
-		const editAndDeleteFastenings = state.fastenings[roomId][stanzaId].filter(
-			(fastening) =>
-				fastening.action === FasteningAction.EDIT || fastening.action === FasteningAction.DELETE
-		);
-		return last(editAndDeleteFastenings);
-	}
-	return undefined;
-};
-
-const filteredReactions: MessageFastening[] = [];
-
-export const getReactionFastenings = (
-	state: RootStore,
-	roomId: string,
-	stanzaId: string
-): MessageFastening[] => {
-	filteredReactions.length = 0;
-	const fastenings = state.fastenings?.[roomId]?.[stanzaId];
-	if (fastenings) {
-		const reactions = filter(
-			fastenings,
-			(fastening) => fastening.action === FasteningAction.REACTION
-		);
-		const latestReactions = reduce(
-			reactions,
-			(acc: { [from: string]: number }, fastening: MessageFastening, index: number) => {
-				if (!acc[fastening.from] || reactions[acc[fastening.from]].date < fastening.date) {
-					acc[fastening.from] = index;
-				}
-				return acc;
-			},
-			{}
-		);
-		filteredReactions.push(
-			...filter(reactions, (_, index) => includes(Object.values(latestReactions), index))
-		);
-		return filteredReactions;
-	}
-	return filteredReactions;
-};
-
-export const getMyLastReaction = (
-	state: RootStore,
-	roomId: string,
-	stanzaId: string
-): string | undefined => {
-	const fastenings = state.fastenings?.[roomId]?.[stanzaId];
-	if (fastenings) {
-		const myReactions = filter(
-			fastenings,
-			(fastening) =>
-				fastening.action === FasteningAction.REACTION && fastening.from === state.session?.id
-		);
-		return last(myReactions)?.value;
-	}
-	return undefined;
-};
