--- src/store/selectors/MarkersSelectors.ts
@@ -1,20 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 Zextras <https://www.zextras.com>
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-
-import { Marker, RoomMarkers } from '../../types/store/MarkersTypes';
-import { RootStore } from '../../types/store/StoreTypes';
-
-export const getMyLastMarkerOfRoom = (store: RootStore, roomId: string): Marker | null => {
-	if (store.session.id && store.markers[roomId] && store.markers[roomId][store.session.id]) {
-		return store.markers[roomId][store.session.id];
-	}
-	return null;
-};
-
-export const getRoomHasMarkers = (store: RootStore, roomId: string): boolean =>
-	!!store.markers[roomId];
-
-export const getMarkers = (store: RootStore, roomId: string): RoomMarkers => store.markers[roomId];
